python scripts: Move more non bitbox-specific modules to communication.
