+++ py/bitbox02/Makefile
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-OUT_DIR=bitbox02/generated
+OUT_DIR=communication/generated
 
 PROTO_FILES=hww.proto common.proto backup_commands.proto bitbox02_system.proto bitboxbase.proto btc.proto eth.proto mnemonic.proto random_number.proto system.proto perform_attestation.proto
 
@@ -33,8 +33,8 @@ ${TMP_TARGETS}: $(addprefix ../../messages/,${PROTO_FILES})
 	touch ${@D}/__init__.py
 # The hack below is due to protobuf-python generating absolute paths (which are frowned upon) instead of relative paths.
 # https://github.com/protocolbuffers/protobuf/issues/1491
-	sed -i 's/^from \([^.]*_pb2\)/from .\1/g'  bitbox02/generated/*.pyi
-	sed -i 's/^import \([^.]*_pb2\) as \([^.]*__pb2\)/from . import \1 as \2/g' bitbox02/generated/*.py
+	sed -i 's/^from \([^.]*_pb2\)/from .\1/g'  ${OUT_DIR}/*.pyi
+	sed -i 's/^import \([^.]*_pb2\) as \([^.]*__pb2\)/from . import \1 as \2/g' ${OUT_DIR}/*.py
 
 release-test: all
 	rm -rf dist/*
@@ -45,3 +45,6 @@ release-live: all
 	rm -rf dist/*
 	python3 setup.py sdist bdist_wheel
 	python3 -m twine upload --sign dist/*
+
+clean:
+	rm -f ${TARGETS}
