+++ py/bitbox02/bitbox02/__init__.py
@@ -53,21 +53,3 @@ from .bitbox02 import (
     UserAbortException,
 )
 from .bootloader import Bootloader
-from .devices import (
-    DeviceInfo,
-    get_any_bitbox02,
-    get_any_bitbox02s,
-    get_any_bitbox02_bootloader,
-    get_any_bitbox02_bootloaders,
-    get_bitbox02multi_bootloader,
-    get_bitbox02multi_bootloaders,
-    get_bitbox02multi_device,
-    get_bitbox02multi_devices,
-    get_bitbox02btc_bootloader,
-    get_bitbox02btc_bootloaders,
-    get_bitbox02btc_device,
-    get_bitbox02btc_devices,
-    parse_device_version,
-    TooManyFoundException,
-    NoneFoundException,
-)
